Scene Log Entries
Creating and incident will now create an incident log entry and any
scene entries associated with that incident.  I am still working on
getting it to recognize when a new scene is created after the initial
incident creation.  It is in the Incident Model where all this logic is
happening.  The incident show method and view is where the log entries
will be displayed.